+++ generator/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased changes
 
+- Update `rust-sdk` so that project is forward-compatible.
 - Updated generator with `plt` and `create-plt` arguments
 
 ## 1.2.0